requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2024.2.2
distro==1.9.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
openai==1.7.2
pydantic==2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.0
python-telegram-bot==20.7
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
annotated-types==0.6.0
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
atlassian-python-api==3.39.0
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
Brotli==1.1.0
bs4==0.0.1
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
comm==0.1.3
comtypes==1.2.1
contourpy==1.2.1
coverage==7.2.7
cssselect==1.2.0
cycler==0.12.1
dateparser==1.2.0
debugpy==1.6.7
decorator==5.1.1
deepdiff==6.3.0
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
dnspython==2.4.2
email_validator==2.1.1
et-xmlfile==1.1.0
executing==1.2.0
fake-useragent==1.1.3
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.16.3
fonttools==4.51.0
fqdn==1.5.1
furl==2.1.3
grpcio==1.60.0
gTTS==2.5.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.3
httptools==0.6.1
httpx==0.26.0
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.5
lxml==4.9.2
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
multidimensional-urlencode==0.0.4
mutagen==1.47.0
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
nose==1.3.7
notebook==6.5.4
notebook_shim==0.2.2
numpy==1.24.2
oauthlib==3.2.2
openai==1.8.0
openpyxl==3.1.2
ordered-set==4.1.0
orderedmultidict==1.0.1
orjson==3.10.3
outcome==1.3.0.post0
packaging==23.1
pandas==2.0.0
pandocfilters==1.5.0
parse==1.19.0
parso==0.8.3
pickleshare==0.7.5
pillow==10.3.0
pip-system-certs==4.0
platformdirs==3.2.0
pook==1.1.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.4
pure-eval==0.2.2
pycparser==2.21
pycryptodomex==3.20.0
pydantic==2.5.3
pydantic_core==2.14.6
pyee==8.2.2
Pygments==2.15.0
pymongo==4.6.1
pyparsing==3.1.2
pypiwin32==223
pyppeteer==1.0.2
pyquery==2.0.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
python-telegram-bot==20.8
pyttsx3==2.90
pytz==2023.3
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
regex==2023.12.25
requests==2.32.3
requests-html==0.10.0
requests-oauthlib==1.3.1
retailcrm==5.1.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
selenium==4.24.0
Send2Trash==1.8.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
stack-data==0.6.2
starlette==0.37.2
terminado==0.17.1
tinycss2==1.2.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.26.2
trio-websocket==0.11.1
typer==0.12.3
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
ujson==5.10.0
uri-template==1.2.0
urllib3==2.0.0
uvicorn==0.29.0
w3lib==2.1.1
watchfiles==0.21.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0
widgetsnbextension==4.0.7
wrapt==1.15.0
wsproto==1.2.0
xlrd==2.0.1
xmltodict==0.13.0
yt-dlp==2024.8.6
zipp==3.15.0
